ansible_collections/junipernetworks/apstra/requirements.txt
ansible-core==2.15.11
aos-sdk @ https://svl-artifactory.juniper.net/artifactory/atom-generic-virtual/aos_sdk_5.0.0-RC1/aos_sdk-0.1.0-py3-none-any.whl#sha256=9f80a035d86de3db68939bbc629cf065db079bfac7565cdf181609b7b763e06b
blinker==1.8.2
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.0
Flask==3.0.3
idna==3.7
iso8601==1.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
lollipop==1.1.8
lollipop-jsonschema==0.8.2
lru-dict==1.3.0
MarkupSafe==2.1.5
netaddr==0.10.1
packaging==24.1
psutil==5.9.8
pycparser==2.22
python-slugify==6.1.2
PyYAML==6.0.2
requests==2.32.3
resolvelib==1.0.1
six==1.16.0
text-unidecode==1.3
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.3
